deleteSub

+++ DoublyLinkedList.java
@@ -1,4 +1,4 @@
-import java.util.NoSuchElementException; 
+import java.util.NoSuchElementException;
 
 
 public class DoublyLinkedList<T extends Comparable<T>> {
@@ -11,7 +11,7 @@ public class DoublyLinkedList<T extends Comparable<T>> {
      * This constructor is used to create a new instance of the DoublyLinkedList
      * class.
      * The head variable will be updated when elements are added to the list.
-     * 
+     *
      * @param <T> the type of elements held in this DoublyLinkedlist
      */
     DoublyLinkedList() {
@@ -20,8 +20,8 @@ public class DoublyLinkedList<T extends Comparable<T>> {
 
     /**
      * Inserts an item into the linked list in its sorted position.
-     * 
-     * @param item the item to be inserted 
+     *
+     * @param item the item to be inserted
      */
     public void insertItem(T item) {
         NodeType<T> newNode = new NodeType<>();
@@ -31,7 +31,7 @@ public class DoublyLinkedList<T extends Comparable<T>> {
         // if list is empty
         if (head == null) {
             head = newNode;
-        } else if (newNode.info.compareTo(head.info) < 0) // insert at the beginning 
+        } else if (newNode.info.compareTo(head.info) < 0) // insert at the beginning
         {
             newNode.next = head;
             head.back = newNode;
@@ -60,44 +60,44 @@ public class DoublyLinkedList<T extends Comparable<T>> {
     }
 
     /**
-     * Deletes the specified item from the linked list 
-     * 
-     * @param item the item to be deleted 
+     * Deletes the specified item from the linked list
+     *
+     * @param item the item to be deleted
      * @throws NoSuchElementException if the item is not found in the linked list.
      */
-    public void deleteItem(T item) 
+    public void deleteItem(T item)
     {
         NodeType<T> temp = searchItem(item);
-      
+
         if (length() == 1)
         {
-            head = null; 
+            head = null;
 } else if (temp.back == null)  // deleting the first item
         {
             NodeType<T> front = temp.next;
             front.back = null;
-            head = front; 
+            head = front;
         } else if ( temp.next == null) // delete last item
         {
             NodeType<T> back = temp.back;
-            back.next = null; 
+            back.next = null;
         } else // delete in the middle
         {
             NodeType<T> front = temp.next;
             NodeType<T> back = temp.back;
             front.back = back;
-            back.next = front; 
+            back.next = front;
         }
     }
 
     /**
      * Searches for an item in the linked list
-     * 
-     * @param item the item is searcing for 
+     *
+     * @param item the item is searcing for
      * @return the node containing the item if found \
-     * @throws NoSuchElementException if the item is not found in the linked list. 
+     * @throws NoSuchElementException if the item is not found in the linked list.
      */
-    public NodeType<T> searchItem(T item) 
+    public NodeType<T> searchItem(T item)
     {
         NodeType<T> temp = head;
         while (temp != null) {
@@ -111,13 +111,13 @@ public class DoublyLinkedList<T extends Comparable<T>> {
         }
         return temp;
     }
-    
+
     /**
      *  Returns the number of elements in the linked list
-     * 
-     * @return the number of elements in the linked list 
+     *
+     * @return the number of elements in the linked list
      */
-    public int length() 
+    public int length()
     {
         int count = 0;
         NodeType<T> temp = head;
@@ -127,12 +127,12 @@ public class DoublyLinkedList<T extends Comparable<T>> {
         }
         return count;
     }
-    
+
     /**
      * Prints the elements of the linked list to the console.
      * The elements are seperated by a spaced and are printed in the order
      * they appear in the list.
-     * A line break is added after printing out all the elements. 
+     * A line break is added after printing out all the elements.
      */
     public void print()
     {
@@ -143,14 +143,14 @@ public class DoublyLinkedList<T extends Comparable<T>> {
         }
         System.out.println();
     }
-    
+
     /**
      * Prints the elements of the linked list in reverse order to the console.
      * The elements are seperated by a space and are printed in reverse order compared
      * to their original order in the list.
      * A line brealk is added after printing all the elements.
      */
-    public void printReverse() 
+    public void printReverse()
     {
         NodeType<T> temp = head;
         while (temp.next != null) {
@@ -162,41 +162,40 @@ public class DoublyLinkedList<T extends Comparable<T>> {
         }
         System.out.println();
     }
-    
+
     /**
      * This function will take input from the user fro the lower
      * and upper bound (both inclusive) for a range of values that
      * you will delete from the list.
-     * 
+     *
      * @param upperBound
      * @param lowerBound
      */
-    public void deleteSubsection(T upperBound, T lowerBound) 
+    public void deleteSubsection(T upperBound, T lowerBound)
     {
         NodeType<T> temp = head;
         while (temp != null) {
-            if ((temp.info.compareTo(lowerBound) >= 0) || temp.info.compareTo(upperBound) <= 0) {
+            if ((temp.info.compareTo(lowerBound) >= 0) && temp.info.compareTo(upperBound) <= 0) {
                 deleteItem(temp.info);
             }
             temp = temp.next;
         }
     }
-    
+
     public void reverseList()
     {
         NodeType<T> prev = null;
-        
-        while (head != null) 
+
+        while (head != null)
         {
             head.next = prev;
             head.back = head.next;
             prev = head;
-            head = head.next; 
+            head = head.next;
         }
     }
 
-    public void swapAlternate() 
-    {
-
+    public void swapAlternate() {
+        while (head
     }
 }
